+++ spec/ig_markets/request_printer_spec.rb
@@ -7,7 +7,7 @@ describe IGMarkets::RequestPrinter do
     IGMarkets::RequestPrinter.enabled = false
   end
 
-  it 'prints options' do
+  it 'prints a request' do
     expect do
       options = {
         method: :get,
@@ -16,35 +16,48 @@ describe IGMarkets::RequestPrinter do
         body: { data: 'ABC' }.to_json
       }
 
-      IGMarkets::RequestPrinter.print_options options
+      IGMarkets::RequestPrinter.print_request options
     end.to output(<<-END
 GET url
   Headers:
     Name: value
     _method: DELETE
-  Body: {
-          "data": "ABC"
-        }
+  Body:
+    {
+      "data": "ABC"
+    }
 END
                  ).to_stdout
   end
 
-  it 'prints a JSON response body' do
+  it 'prints a JSON response' do
+    response = instance_double 'Excon::Response', headers: { 'Name' => 'Value' }, body: { data: 'ABC' }.to_json
+
     expect do
-      IGMarkets::RequestPrinter.print_response_body({ data: 'ABC' }.to_json)
+      IGMarkets::RequestPrinter.print_response response
     end.to output(<<-END
-  Response: {
-              "data": "ABC"
-            }
+  Response:
+    Headers:
+      Name: Value
+    Body:
+      {
+        "data": "ABC"
+      }
 END
                  ).to_stdout
   end
 
-  it 'prints an HTML response body' do
+  it 'prints an HTML response' do
+    response = instance_double 'Excon::Response', headers: { 'Name' => 'Value' }, body: '<html></html>'
+
     expect do
-      IGMarkets::RequestPrinter.print_response_body '<html></html>'
+      IGMarkets::RequestPrinter.print_response response
     end.to output(<<-END
-  Response: <html></html>
+  Response:
+    Headers:
+      Name: Value
+    Body:
+      <html></html>
 END
                  ).to_stdout
   end